The program at least sort of works
It runs with no arguments!
